requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
supabase==2.11.0
pydantic==2.10.5
pytest==8.3.4
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0